Move project to new structure
