+++ internal/zbx/node_status_test.go
